request camera permission before DrawAR
