+++ app/src/main/AndroidManifest.xml
@@ -17,28 +17,47 @@
     package="com.googlecreativelab.drawar">
 
     <uses-permission android:name="android.permission.CAMERA" />
-    <!-- This tag indicates that this application requires ARCore.  This results in the application
-         only being visible in the Google Play Store on devices that support ARCore. -->
-    <uses-feature android:name="android.hardware.camera.ar" android:required="true"/>
+    <!--
+         This tag indicates that this application requires ARCore.  This results in the application
+         only being visible in the Google Play Store on devices that support ARCore.
+    -->
+    <uses-feature
+        android:name="android.hardware.camera.ar"
+        android:required="true" />
+
     <application
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/AppTheme" >
-        <activity
-            android:name=".DrawAR"
+        android:theme="@style/AppTheme">
+
+        <activity android:name=".FakeMainActivity"
             android:configChanges="orientation|screenSize"
             android:label="@string/app_name"
             android:screenOrientation="locked"
             android:theme="@style/Theme.AppCompat.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <!-- This tag indicates that this application requires ARCore.  This results in the Google Play
-     Store downloading and installing ARCore along with the application. -->
-        <meta-data android:name="com.google.ar.core" android:value="required" />
+
+        <activity
+            android:name=".DrawAR"
+            android:configChanges="orientation|screenSize"
+            android:label="@string/app_name"
+            android:screenOrientation="locked"
+            android:theme="@style/Theme.AppCompat.NoActionBar">
+
+        </activity>
+        <!--
+     This tag indicates that this application requires ARCore.  This results in the Google Play
+     Store downloading and installing ARCore along with the application.
+        -->
+        <meta-data
+            android:name="com.google.ar.core"
+            android:value="required" />
     </application>
 
 </manifest>
